backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
newspaper3k==0.2.8
selenium==4.16.0

# AI & LangChain
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0
tiktoken==0.5.2

# Task Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
